+++ psi4/src/psi4/libmints/fjt.cc
@@ -53,13 +53,17 @@
 // The U.S. Government is granted a limited license as per AL 91-7.
 //
 
-#include <cmath>
 #include "integral.h"
 #include "fjt.h"
 #include "wavefunction.h"
 #include "integralparameters.h"
+
 #include "psi4/libciomr/libciomr.h"
 #include "psi4/psi4-dec.h"
+#include "psi4/libpsi4util/PsiOutStream.h"
+
+#include <cmath>
+
 using namespace psi;
 using namespace std;
 ;
@@ -85,7 +89,7 @@ double Taylor_Fjt::relative_zero_(1e-6);
   Initialize Taylor_Fm_Eval object (computes incomplete
   gamma function via Taylor interpolation)
  ------------------------------------------------------*/
-Taylor_Fjt::Taylor_Fjt(unsigned int mmax, double accuracy) :
+Taylor_Fjt::Taylor_Fjt(size_t mmax, double accuracy) :
     cutoff_(accuracy), interp_order_(TAYLOR_INTERPOLATION_ORDER),
     F_(new double[mmax+1])
 {
